fix: harden migrations and runner (quality review fixes)
- Add CASCADE to all 13 down migrations for deterministic rollbacks
- Add ON DELETE RESTRICT explicitly on billing_invoices/payments tenant FK
- Change monetary INTEGER columns to BIGINT (price_cents, amount_cents, discount_value, limit_value)
- Add updated_at to coupons; CHECK constraint on tenant_feature_overrides
- Add missing indexes: billing_payments(invoice_id, tenant+status), tenant_subscriptions(tenant_id), audit_logs(tenant+created_at), subscription_plans(active)
- Fix sslmode=disable hardcoded in migrate runner; read POSTGRES_SSLMODE from env (default: require)
- Validate required env vars before building DSN
- Document POSTGRES_SSLMODE=disable in .env.example for local dev

+++ services/api/migrations/000007_create_plan_features.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE plan_features (
     plan_id       UUID NOT NULL REFERENCES subscription_plans(id) ON DELETE CASCADE,
     feature_key   VARCHAR(128) NOT NULL,
     enabled       BOOLEAN NOT NULL DEFAULT TRUE,
-    limit_value   INTEGER,
+    limit_value   BIGINT,
     trial_enabled BOOLEAN NOT NULL DEFAULT FALSE,
     trial_days    INTEGER,
     metadata_json JSONB,
